requirements.txt

# Core Framework
# Document Processing
# Embeddings
# LLM & Agents
# Utilities
# Vector Database
aiofiles>=23.0.0
fastapi>=0.115.0,<0.124.0
google-adk>=1.0.0
ollama>=0.3.0
openpyxl>=3.1.0
pandas>=2.0.0
pydantic>=2.0.0
pymilvus>=2.4.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
python-pptx>=1.0.0
reportlab>=4.0.0
sentence-transformers>=3.0.0
streamlit>=1.30.0
tabulate
tabulate>=0.9.0
uvicorn[standard]>=0.27.0
